pyproject.toml
[tool.poetry]
name = "pytools"
version = "0.1.0"
description = "some python utility scripts"
authors = ["lalimy"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
requests = "^2.27.1"
requirements-parser = "^0.5.0"
attrs = "^20.3.0"
colorama = "^0.4.6"
pytest-cov = "^3.0.0"
pytest = "^7.1.2"
pytest-dependency = "^0.5.1"


[tool.poetry.dev-dependencies]

[settings.virtualenvs.in-project]

[build-system]
requires = ["poetry>=1.1.2"]
build-backend = "poetry.masonry.api"
